+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/spi/combined/CombinedGenerators.java
@@ -6,7 +6,6 @@ import io.github.ulisse1996.jaorm.spi.GeneratorsService;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class CombinedGenerators extends GeneratorsService {
 
@@ -15,7 +14,7 @@ public class CombinedGenerators extends GeneratorsService {
     public CombinedGenerators(List<GeneratorsService> services) {
         this.generated = services.stream()
                 .flatMap(s -> s.getGenerated().entrySet().stream())
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+                .collect(Combiners.getEntryMapCollector());
     }
 
     @Override