+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/spi/combined/CombinedQueries.java
@@ -6,7 +6,6 @@ import io.github.ulisse1996.jaorm.spi.QueriesService;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class CombinedQueries extends QueriesService {
 
@@ -15,7 +14,7 @@ public class CombinedQueries extends QueriesService {
     public CombinedQueries(List<QueriesService> services) {
         this.queries = services.stream()
                 .flatMap(s -> s.getQueries().entrySet().stream())
-                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+                .collect(Combiners.getEntryMapCollector());
     }
 
     @Override