+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/exception/ProcessorException.java
@@ -5,4 +5,8 @@ public class ProcessorException extends RuntimeException {
     public ProcessorException(String exception) {
         super(exception);
     }
+
+    public ProcessorException(String message, Throwable ex) {
+        super(message, ex);
+    }
 }