+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/generation/GenerationType.java
@@ -9,7 +9,6 @@ public enum GenerationType {
     CONVERTER,
     LISTENERS,
     GENERATORS,
-    TABLES,
     PROJECTIONS,
     GRAPH
 }