+++ jaorm-test/src/test/java/io/github/ulisse1996/jaorm/integration/test/CoreIT.java
@@ -5,7 +5,6 @@ import io.github.ulisse1996.jaorm.entity.EntityComparator;
 import io.github.ulisse1996.jaorm.entity.Result;
 import io.github.ulisse1996.jaorm.exception.JaormSqlException;
 import io.github.ulisse1996.jaorm.exception.JaormValidationException;
-import io.github.ulisse1996.jaorm.generated.Tables;
 import io.github.ulisse1996.jaorm.integration.test.entity.*;
 import io.github.ulisse1996.jaorm.integration.test.projection.MyProjection;
 import io.github.ulisse1996.jaorm.integration.test.projection.ProjectionDao;
@@ -180,30 +179,6 @@ class CoreIT extends AbstractIT {
 
         Assertions.assertEquals(BigDecimal.ONE, generated.getId());
         Assertions.assertEquals(BigDecimal.valueOf(2), generated.getSeq());
-    }
-
-    @ParameterizedTest
-    @MethodSource("getSqlTests")
-    void should_read_entity_using_tables(HSQLDBProvider.DatabaseType type, String initSql) {
-        setDataSource(type, initSql);
-
-        Optional<City> cityOpt = Tables.CITY.cityId(10)
-                .readOpt();
-
-        City city = new City();
-        city.setCityId(11);
-        city.setName("NAME");
-
-
-        Assertions.assertFalse(cityOpt.isPresent());
-
-        QueriesService.getInstance().getQuery(CityDAO.class).insert(city);
-
-        cityOpt = Tables.CITY.cityId(11)
-                .readOpt();
-
-        Assertions.assertTrue(cityOpt.isPresent());
-        Assertions.assertTrue(EntityComparator.getInstance(City.class).equals(city, cityOpt.get()));
     }
 
     @ParameterizedTest